==> 4_ Final_Score_Calculation.ipynb <==
# Import all relevant libraries to the note book
import osmnx as ox
import pandas as pd
import networkx as nx
import numpy as np
import geopandas as gpd
import matplotlib.pyplot as plt
import shapely
import fiona
import jenkspy

Set Up GeoDataFrame

# Load the walking network with the common code
area_shape = fiona.open("root/Inputs/GemeenteAmsterdam.geojson")
area_shape = shapely.geometry.shape(area_shape[0]['geometry'])

cf = """
     ["area"!~"yes"]
     ["highway"]
     ["highway"!~"motor|proposed|construction|abandoned|platform|raceway"]
     ["foot"!~"no"]
     ["service"!~"private"]
     ["access"!~"private"]
     """

G = ox.graph_from_polygon(polygon=area_shape, custom_filter=cf, simplify=False, retain_all=True, network_type="walk", truncate_by_edge=True)

# Convert to GeoDataFrame
geoDataFrame = ox.graph_to_gdfs(G, nodes=False, edges=True,  node_geometry=False, fill_edge_geometry=True)

#Ensure that the u column is not a index so reset accordingly
geoDataFrame = geoDataFrame.reset_index()
print(geoDataFrame)

# To prevent double counting of two way streets the edges are grouped by the same u and v columns and then split again after selecting for the first street
geoDataFrame['edge'] = geoDataFrame.apply(lambda row: tuple(sorted([row['u'], row['v']])), axis=1)
gdf_combined = geoDataFrame.groupby('edge').first().reset_index()
gdf_combined[['u', 'v']] = pd.DataFrame(gdf_combined['edge'].tolist(), index=gdf_combined.index)

print(gdf_combined)

#Store network in case it is needed
with open('./root/Outputs/Network_Amsterdam.geojson' , 'w') as file:
    file.write(gdf_combined.to_json())

#Pull all files from previous script - should be geojson or geo-package
file_mobility = "./root/Outputs/2_modelled_pedestrian_intensity.gpkg" # fill in full path
file_shade_PET = "./root/Outputs/3_Shade+PET.gpkg" # fill in full path

#Read the files into GeoDataFrames and save the columns that are needed only
gdf_mobility = gpd.read_file(file_mobility)
gdf_mobility = gdf_mobility[['u', 'v', 'pop', 'buurtcode', 'young_pop_pct', 'old_pop_pct', 'pop_unknown_pct']]
gdf_mobility['u'] = gdf_mobility['u'].astype('int64')
gdf_mobility['v'] = gdf_mobility['v'].astype('int64')
gdf_shade = gpd.read_file(file_shade_PET)
gdf_shade = gdf_shade[['u', 'v', "0900",  "0930", "1000", "1030", "1100", "1130", "1200", "1230", "1300", "1330", "1400", "1430", "1500", "1530", "1600", "1630", "1700", "1730", "1800", "1830", "1900", "1930", "2000", "sum_adjust", 'PET', "avg_exposure_percent", 'avg_shade_percent']]

#Again to prevent double counting the first street is selected
gdf_shade['edge'] = gdf_shade.apply(lambda row: tuple(sorted([row['u'], row['v']])), axis=1)
gdf_shade_combined = gdf_shade.groupby('edge').first().reset_index()
gdf_shade_combined [['u', 'v']] = pd.DataFrame(gdf_combined['edge'].tolist(), index=gdf_combined.index)

print(gdf_shade_combined)
print(gdf_mobility)
print(gdf_combined)

#Combine all GeoDataFrames into one file
gdf_heat_combined = gdf_combined.merge(gdf_shade_combined, how = 'left', on =['u', 'v']) 
gdf_heat_combined = gdf_heat_combined.merge(gdf_mobility, how = 'left', on =['u', 'v']) 
print(gdf_heat_combined.head(30))

Set Geometry

#Set the geometry as it is defined by the first network pulled for the GeoDataFrame
gdf_heat_combined = gdf_heat_combined.set_geometry(gdf_heat_combined["geometry"], inplace=False)
gdf_heat_combined = gdf_heat_combined.fillna(0)

Generate threshold functions

#Fill in strings for column names
#Funtion which loops through various scores and sets the column value to 1 or 0 accordingly for threshold functionality
def shade_treshold(value):
    if value < 0.4:
        return 1
    else:
        return 0

def PET_treshold(value):
    if value >= 35:
        return 1
    else:
        return 0
    
def Business_bound(value):
    if value != 0:
        return 1
    else:
        return 0

def top_percent(value):
    if value > 0.96:
        return 1
    else:
        return 0

Do the math

#Function calculates our score based on the columns from the GeoDataFrame
def calculate_score(gdf_combined, column1_shade_sum, column2_PET, column3_mob, column1_shade_perc):
    max_shade = gdf_combined[column1_shade_sum].max()
    min_shade = gdf_combined[column1_shade_sum].min()
    max_PET = gdf_combined[column2_PET].max()
    min_PET = gdf_combined[column2_PET].min()


    #Set column to 0 or 1 based on threshold values above
    gdf_combined['shade_tresh'] = gdf_combined[column1_shade_perc].apply(shade_treshold)
    gdf_combined['PET_tresh'] = gdf_combined[column2_PET].apply(PET_treshold)

    #Reprojected Shade and PET score from 0-1 to make them comparable
    gdf_combined['shade_score'] = ((gdf_combined[column1_shade_sum]-min_shade)/(max_shade-min_shade)) 
    gdf_combined['PET_score'] = ((gdf_combined[column2_PET]-min_PET)/(max_PET-min_PET))

    #Calculate final risk score per street
    gdf_combined['Final_score_all'] = gdf_combined['shade_tresh']*gdf_combined['PET_tresh']*(((gdf_combined['shade_score']) + (gdf_combined['PET_score']))/2) #Add weights to criteria if wanted
    
    #Calulate top 4% percent of streets in Amsterdam based on modelled intensity, requirement that Final_score_all is not 0
    gdf_combined['busy_select'] = (gdf_combined['Final_score_all'].apply(Business_bound))
    gdf_combined['busy_after_crit'] = gdf_combined[column3_mob]*gdf_combined['busy_select']
    gdf_combined['busy_after_crit'] = gdf_combined['busy_after_crit'].rank(pct = True)
    gdf_combined['busy_treshold'] =  gdf_combined['busy_after_crit'].apply(top_percent)
    gdf_combined['Final_score_4_perc'] = gdf_combined['Final_score_all']*gdf_combined['busy_treshold']

    #Add column defining with category when displaying the flow values later
    gdf_jenks = gdf_combined.loc[gdf_combined[column3_mob] != 0]
    jnb = jenkspy.JenksNaturalBreaks(3)
    jnb.fit(gdf_jenks[column3_mob])
    brea = jnb.breaks_
    breaki = [0] + brea
    gdf_combined['jenkins_bin'] = pd.cut(gdf_combined[column3_mob], bins=breaki, labels=['bin_1', 'bin_2', 'bin_3', 'bin_4'], include_lowest = True) 
    
    return gdf_combined

#Implementing the previous function
gdf_heat_combined = calculate_score(gdf_heat_combined, 'sum_adjust', 'PET', 'pop', 'avg_shade_percent')
gdf_heat_combined = gdf_heat_combined.round(decimals=3)

print(gdf_heat_combined.head(30))

Export as GEOJson

#Export file as GEOJson
with open('./root/Outputs/4_Final_Heat_Risk_Score.geojson' , 'w') as file:
    file.write(gdf_heat_combined.to_json())
